feat: Move react-virtual into vendored dependencies
