+++ .gitignore
@@ -9,4 +9,4 @@ src/index.ts
 src/test-utils/dom/index.ts
 src/test-utils/selectors
 src/internal/generated/custom-css-properties/index.*
-vendor/generated-third-party-licenses.txt
+vendor/generated-*.txt
