+++ package.json
@@ -43,7 +43,6 @@
     "mnth": "^2.0.0",
     "react-keyed-flatten-children": "^1.3.0",
     "react-transition-group": "^4.4.2",
-    "react-virtual": "^2.8.2",
     "tslib": "^2.4.0",
     "weekstart": "^1.1.0"
   },
@@ -120,6 +119,7 @@
     "react-dom": "^16.14.0",
     "react-router-dom": "^5.3.0",
     "react-test-renderer": "^16.14.0",
+    "react-virtual": "^2.8.2",
     "replace-in-files": "^3.0.0",
     "rimraf": "^5.0.5",
     "rollup": "^3.5.1",
